change support chield (orders) mechanism; count backers, collected and remainder

+++ db/migrate/20160901143957_add_contribution_backers_remainder_to_products.rb
@@ -1,6 +1,6 @@
 class AddContributionBackersRemainderToProducts < ActiveRecord::Migration
   def change
-    add_column :products, :contribution, :integer, default: 0
+    add_column :products, :contribution, :decimal, default: 0
     add_column :products, :backers, :integer, default: 0
     add_column :products, :remainder, :integer, default: 0
   end
